chore: made suggested PR changes, consolidated const and type definitions
