+++ src/flows/pipes/Visualization/style.scss
@@ -18,6 +18,10 @@
   .empty-graph-error {
     height: 200px;
   }
+
+  .qr-code {
+    width: 60%;
+  }
 }
 .flow-visualization--view-error {
   background-color: $c-fire;
